refactor: WI-0657 leave promotion delivery history core helper extraction
