+++ ROADMAP.md
@@ -938,3 +938,4 @@ Phase 8: Extensions (ATS, performance, expenses, analytics)
 - WI-0654 withholding copy-runtime line-budget recovery (reduce src/components/withholding-receipt/copy-runtime.ts to stay within <=380 lines guard without behavior changes and add e2e-wi0654 regression guard)
 - WI-0655 KO surface one-shot sweep and CI guard (expand KO guard coverage for withholding/payslips/contracts locale/runtime copy blocks, lock contract fallback strings + withholding activity fallback, and chain WI-0522/WI-0579/WI-0655 under test:e2e:ko-guard)
 - WI-0656 scheduling rotation assignment core helper extraction (extract shared rotation assignment helpers for template resolution, weekday-set validation, overlap preflight, and schedule creation into rotation-assignment-core-helpers.ts; rewire service call sites with callback injection; add e2e-wi0656 guard; keep scheduling/service.ts <=2900 lines)
+- WI-0657 leave promotion delivery history core helper extraction (extract recipient stats, delivery-history persistence, and retry recipient-id normalization into promotion-delivery-history-core-helpers.ts; rewire dispatch/retry call sites; add e2e-wi0657 guard; keep leave/service.ts <=2450 lines)
