+++ scripts/tests/e2e-wi0461-leave-promotion-history-view-helper-extraction-line-budget-2850.test.ts
@@ -29,7 +29,6 @@ async function run() {
   assert.match(leaveService, /toPromotionDeliverySummaryView/);
   assert.match(leaveService, /toPromotionDeliveryRecipientView/);
   assert.match(leaveService, /toRetryCountByEmployeeId/);
-  assert.match(leaveService, /toRecipientStatus/);
 
   assert.doesNotMatch(leaveService, /function toPromotionTargetSnapshots\(/);
   assert.doesNotMatch(leaveService, /function toPromotionDeliverySummaryView\(/);
